1. update parse_sentinel_xml to accormodate different granule names. 2. read Sentinel-2 scene classification SCL as the SR QA flag to differentiate cloud and snow. 3. update a lot of changes to data types from int8 to uint16 due to the change of files for QA flag. 4. enhance getting projection info from xml, all GTCP projection codes are added and can be dynamically chosen. 5. update on the shell script to run sentinel albedo 6. misc change to the shell script for query of sentinel-2

+++ sentinel-2-albedo/albedo/lndAlbedo_main.c
@@ -116,11 +116,11 @@ int main(int argc, char *argv[])
   
    //assert(0 == subset_modis(sensor, modis));
 
-  /* retrieve meta data from MODIS surface reflectance */
+  // retrieve meta data from MODIS surface reflectance
   if(getModisMetaInfo(modis)==FAILURE) {
     fprintf(stderr, "\nmain: Retrieve MODIS metadata error\n");
     exit(1);
-  }   
+  }
 
   /*Distance matrix among Classes */
   alloc_2dim_contig((void ***)(&(sensor->AmongClsDistance)),  sensor->init_ncls+1,sensor->init_ncls+1, sizeof(double));
@@ -137,8 +137,8 @@ int main(int argc, char *argv[])
   alloc_1dim_contig((void **) (&sensor->Clsdata), sensor->ncols, sizeof(int8));
  
   /* Allocate memory for one Row input lndSR, atmo,lndSRQA*/
- alloc_2dim_contig((void ***)(&(sensor->OneRowlndSR)), SENSOR_BANDS,sensor->ncols, sizeof(int16));
- alloc_1dim_contig((void **)(&(sensor->OneRowlndSRQAData)), sensor->ncols, sizeof(uint8));
+ alloc_2dim_contig((void ***)(&(sensor->OneRowlndSR)), SENSOR_BANDS,sensor->ncols, sizeof(uint16));
+ alloc_1dim_contig((void **)(&(sensor->OneRowlndSRQAData)), sensor->ncols, sizeof(uint16));
  alloc_1dim_contig((void **)(&(sensor->OneRowlndSRAtmoData)), sensor->ncols, sizeof(int16));
  alloc_1dim_contig((void **)(&(sensor->OneRowlndSRCldData)), sensor->ncols, sizeof(uint16));
  alloc_1dim_contig((void **)(&(sensor->OneRowlndSRSnwData)), sensor->ncols, sizeof(uint16));
